upadated data fetching with LINQ

+++ ExordiumApp/Assets/Scripts/SerializableClasses.cs
@@ -10,7 +10,7 @@ public class Item
     public string name;
     public string description;
     public string image_url;
-    public decimal price;
+    public float price;
     public int retailer_id;
     public int item_category_id;
 }
